Get rid of unused components.
