--- src/components/Attribution/index.js
@@ -1,19 +1,0 @@
-import React from 'react'
-
-const Attribution = () => (
-  <p className="attribution">
-    <a href="https://github.com/Uniswap/uniswap-info" rel="noopener noreferrer" target="_blank">
-      Github
-    </a>{' '}
-    |{' '}
-    <a href="https://uniswap.io" rel="noopener noreferrer" target="_blank">
-      Jelly Swap
-    </a>{' '}
-    |{' '}
-    <a href="https://gifer.com/en/1Atv" rel="noopener noreferrer" target="_blank">
-      GIF
-    </a>
-  </p>
-)
-
-export default Attribution
