--- src/components/Emoji/index.js
@@ -1,14 +1,0 @@
-import React from 'react'
-
-const Emoji = props => (
-  <span
-    className="emoji"
-    role="img"
-    aria-label={props.label ? props.label : ''}
-    aria-hidden={props.label ? 'false' : 'true'}
-  >
-    {props.symbol}
-  </span>
-)
-
-export default Emoji
