--- src/contexts/GlobalData.js
@@ -1,632 +1,0 @@
-import React, { createContext, useContext, useReducer, useMemo, useCallback, useEffect, useState } from 'react'
-import { client } from '../apollo/client'
-import axios from 'axios'
-import dayjs from 'dayjs'
-import utc from 'dayjs/plugin/utc'
-import { useTimeframe } from './Application'
-import {
-  getPercentChange,
-  getBlockFromTimestamp,
-  getBlocksFromTimestamps,
-  get2DayPercentChange,
-  getTimeframe
-} from '../utils'
-import { GLOBAL_DATA, GLOBAL_TXNS, ETH_PRICE, ALL_PAIRS, ALL_TOKENS, TOP_LPS_PER_PAIRS } from '../apollo/queries'
-import weekOfYear from 'dayjs/plugin/weekOfYear'
-import { useAllPairData } from './PairData'
-const UPDATE = 'UPDATE'
-const UPDATE_TXNS = 'UPDATE_TXNS'
-const UPDATE_CHART = 'UPDATE_CHART'
-const UPDATE_ETH_PRICE = 'UPDATE_ETH_PRICE'
-const ETH_PRICE_KEY = 'ETH_PRICE_KEY'
-const UPDATE_ALL_PAIRS_IN_UNISWAP = 'UPDAUPDATE_ALL_PAIRS_IN_UNISWAPTE_TOP_PAIRS'
-const UPDATE_ALL_TOKENS_IN_UNISWAP = 'UPDATE_ALL_TOKENS_IN_UNISWAP'
-const UPDATE_TOP_LPS = 'UPDATE_TOP_LPS'
-
-// format dayjs with the libraries that we need
-dayjs.extend(utc)
-dayjs.extend(weekOfYear)
-
-const GlobalDataContext = createContext()
-
-function useGlobalDataContext() {
-  return useContext(GlobalDataContext)
-}
-
-function reducer(state, { type, payload }) {
-  switch (type) {
-    case UPDATE: {
-      const { data } = payload
-      return {
-        ...state,
-        globalData: data
-      }
-    }
-    case UPDATE_TXNS: {
-      const { transactions } = payload
-      return {
-        ...state,
-        transactions
-      }
-    }
-    case UPDATE_CHART: {
-      const { daily, weekly } = payload
-      return {
-        ...state,
-        chartData: {
-          daily,
-          weekly
-        }
-      }
-    }
-    case UPDATE_ETH_PRICE: {
-      const { ethPrice, oneDayPrice, ethPriceChange } = payload
-      return {
-        [ETH_PRICE_KEY]: ethPrice,
-        oneDayPrice,
-        ethPriceChange
-      }
-    }
-
-    case UPDATE_ALL_PAIRS_IN_UNISWAP: {
-      const { allPairs } = payload
-      return {
-        ...state,
-        allPairs
-      }
-    }
-
-    case UPDATE_ALL_TOKENS_IN_UNISWAP: {
-      const { allTokens } = payload
-      return {
-        ...state,
-        allTokens
-      }
-    }
-
-    case UPDATE_TOP_LPS: {
-      const { topLps } = payload
-      return {
-        ...state,
-        topLps
-      }
-    }
-    default: {
-      throw Error(`Unexpected action type in DataContext reducer: '${type}'.`)
-    }
-  }
-}
-
-export default function Provider({ children }) {
-  const [state, dispatch] = useReducer(reducer, {})
-  const update = useCallback(data => {
-    dispatch({
-      type: UPDATE,
-      payload: {
-        data
-      }
-    })
-  }, [])
-
-  const updateTransactions = useCallback(transactions => {
-    dispatch({
-      type: UPDATE_TXNS,
-      payload: {
-        transactions
-      }
-    })
-  }, [])
-
-  const updateChart = useCallback((daily, weekly) => {
-    dispatch({
-      type: UPDATE_CHART,
-      payload: {
-        daily,
-        weekly
-      }
-    })
-  }, [])
-
-  const updateEthPrice = useCallback((ethPrice, oneDayPrice, ethPriceChange) => {
-    dispatch({
-      type: UPDATE_ETH_PRICE,
-      payload: {
-        ethPrice,
-        oneDayPrice,
-        ethPriceChange
-      }
-    })
-  }, [])
-
-  const updateAllPairsInUniswap = useCallback(allPairs => {
-    dispatch({
-      type: UPDATE_ALL_PAIRS_IN_UNISWAP,
-      payload: {
-        allPairs
-      }
-    })
-  }, [])
-
-  const updateAllTokensInUniswap = useCallback(allTokens => {
-    dispatch({
-      type: UPDATE_ALL_TOKENS_IN_UNISWAP,
-      payload: {
-        allTokens
-      }
-    })
-  }, [])
-
-  const updateTopLps = useCallback(topLps => {
-    dispatch({
-      type: UPDATE_TOP_LPS,
-      payload: {
-        topLps
-      }
-    })
-  }, [])
-  return (
-    <GlobalDataContext.Provider
-      value={useMemo(
-        () => [
-          state,
-          {
-            update,
-            updateTransactions,
-            updateChart,
-            updateEthPrice,
-            updateTopLps,
-            updateAllPairsInUniswap,
-            updateAllTokensInUniswap
-          }
-        ],
-        [
-          state,
-          update,
-          updateTransactions,
-          updateTopLps,
-          updateChart,
-          updateEthPrice,
-          updateAllPairsInUniswap,
-          updateAllTokensInUniswap
-        ]
-      )}
-    >
-      {children}
-    </GlobalDataContext.Provider>
-  )
-}
-
-/**
- * Gets all the global data for the overview page.
- * Needs current eth price and the old eth price to get
- * 24 hour USD changes.
- * @param {*} ethPrice
- * @param {*} oldEthPrice
- */
-async function getGlobalData(ethPrice, oldEthPrice) {
-  // data for each day , historic data used for % changes
-  let data = {}
-  let oneDayData = {}
-  let twoDayData = {}
-
-  try {
-    // get timestamps for the days
-    const utcCurrentTime = dayjs()
-    const utcOneDayBack = utcCurrentTime.subtract(1, 'day').unix()
-    const utcTwoDaysBack = utcCurrentTime.subtract(2, 'day').unix()
-    const utcOneWeekBack = utcCurrentTime.subtract(1, 'week').unix()
-    const utcTwoWeeksBack = utcCurrentTime.subtract(2, 'week').unix()
-
-    // get the blocks needed for time travel queries
-    let [oneDayBlock, twoDayBlock, oneWeekBlock, twoWeekBlock] = await getBlocksFromTimestamps([
-      utcOneDayBack,
-      utcTwoDaysBack,
-      utcOneWeekBack,
-      utcTwoWeeksBack
-    ])
-
-    // fetch the global data
-    let result = await client.query({
-      query: GLOBAL_DATA(),
-      fetchPolicy: 'cache-first'
-    })
-    data = result.data.uniswapFactories[0]
-
-    // fetch the historical data
-    let oneDayResult = await client.query({
-      query: GLOBAL_DATA(oneDayBlock?.number),
-      fetchPolicy: 'cache-first'
-    })
-    oneDayData = oneDayResult.data.uniswapFactories[0]
-
-    let twoDayResult = await client.query({
-      query: GLOBAL_DATA(twoDayBlock?.number),
-      fetchPolicy: 'cache-first'
-    })
-    twoDayData = twoDayResult.data.uniswapFactories[0]
-
-    let oneWeekResult = await client.query({
-      query: GLOBAL_DATA(oneWeekBlock?.number),
-      fetchPolicy: 'cache-first'
-    })
-    const oneWeekData = oneWeekResult.data.uniswapFactories[0]
-
-    let twoWeekResult = await client.query({
-      query: GLOBAL_DATA(twoWeekBlock?.number),
-      fetchPolicy: 'cache-first'
-    })
-    const twoWeekData = twoWeekResult.data.uniswapFactories[0]
-
-    if (data && oneDayData && twoDayData && twoWeekData) {
-      let [oneDayVolumeUSD, volumeChangeUSD] = get2DayPercentChange(
-        data.untrackedVolumeUSD,
-        oneDayData.untrackedVolumeUSD ? oneDayData.untrackedVolumeUSD : 0,
-        twoDayData.untrackedVolumeUSD ? twoDayData.untrackedVolumeUSD : 0
-      )
-
-      const [oneWeekVolume, weeklyVolumeChange] = get2DayPercentChange(
-        data.untrackedVolumeUSD,
-        oneWeekData.untrackedVolumeUSD,
-        twoWeekData.untrackedVolumeUSD
-      )
-
-      const [oneDayTxns, txnChange] = get2DayPercentChange(
-        data.txCount,
-        oneDayData.txCount ? oneDayData.txCount : 0,
-        twoDayData.txCount ? twoDayData.txCount : 0
-      )
-
-      // format the total liquidity in USD
-      data.totalLiquidityUSD = data.totalLiquidityETH * ethPrice
-      const liquidityChangeUSD = getPercentChange(
-        data.totalLiquidityETH * ethPrice,
-        oneDayData.totalLiquidityETH * oldEthPrice
-      )
-
-      // add relevant fields with the calculated amounts
-      data.oneDayVolumeUSD = oneDayVolumeUSD
-      data.oneWeekVolume = oneWeekVolume
-      data.weeklyVolumeChange = weeklyVolumeChange
-      data.volumeChangeUSD = volumeChangeUSD
-      data.liquidityChangeUSD = liquidityChangeUSD
-      data.oneDayTxns = oneDayTxns
-      data.txnChange = txnChange
-    }
-  } catch (e) {
-    console.log(e)
-  }
-
-  return data
-}
-
-/**
- * Get and format transactions for global page
- */
-const getGlobalTransactions = async () => {
-  let transactions = {}
-
-  try {
-    let result = await client.query({
-      query: GLOBAL_TXNS,
-      fetchPolicy: 'cache-first'
-    })
-    transactions.mints = []
-    transactions.burns = []
-    transactions.swaps = []
-    result?.data?.transactions &&
-      result.data.transactions.map(transaction => {
-        if (transaction.mints.length > 0) {
-          transaction.mints.map(mint => {
-            return transactions.mints.push(mint)
-          })
-        }
-        if (transaction.burns.length > 0) {
-          transaction.burns.map(burn => {
-            return transactions.burns.push(burn)
-          })
-        }
-        if (transaction.swaps.length > 0) {
-          transaction.swaps.map(swap => {
-            return transactions.swaps.push(swap)
-          })
-        }
-        return true
-      })
-  } catch (e) {
-    console.log(e)
-  }
-
-  return transactions
-}
-
-/**
- * Gets the current price  of ETH, 24 hour price, and % change between them
- */
-const getEthPrice = async () => {
-  const utcCurrentTime = dayjs()
-  const utcOneDayBack = utcCurrentTime
-    .subtract(1, 'day')
-    .startOf('minute')
-    .unix()
-
-  let ethPrice = 0
-  let ethPriceOneDay = 0
-  let priceChangeETH = 0
-
-  try {
-    let oneDayBlock = await getBlockFromTimestamp(utcOneDayBack)
-    let result = await client.query({
-      query: ETH_PRICE(),
-      fetchPolicy: 'cache-first'
-    })
-    let resultOneDay = await client.query({
-      query: ETH_PRICE(oneDayBlock),
-      fetchPolicy: 'cache-first'
-    })
-    const currentPrice = result?.data?.bundles[0]?.ethPrice
-    const oneDayBackPrice = resultOneDay?.data?.bundles[0]?.ethPrice
-    priceChangeETH = getPercentChange(currentPrice, oneDayBackPrice)
-    ethPrice = currentPrice
-    ethPriceOneDay = oneDayBackPrice
-  } catch (e) {
-    console.log(e)
-  }
-
-  return [ethPrice, ethPriceOneDay, priceChangeETH]
-}
-
-const PAIRS_TO_FETCH = 500
-const TOKENS_TO_FETCH = 500
-
-/**
- * Loop through every pair on uniswap, used for search
- */
-async function getAllPairsOnUniswap() {
-  try {
-    let allFound = false
-    let pairs = []
-    let skipCount = 0
-    while (!allFound) {
-      let result = await client.query({
-        query: ALL_PAIRS,
-        variables: {
-          skip: skipCount
-        },
-        fetchPolicy: 'cache-first'
-      })
-      skipCount = skipCount + PAIRS_TO_FETCH
-      pairs = pairs.concat(result?.data?.pairs)
-      if (result?.data?.pairs.length < PAIRS_TO_FETCH || pairs.length > PAIRS_TO_FETCH) {
-        allFound = true
-      }
-    }
-    return pairs
-  } catch (e) {
-    console.log(e)
-  }
-}
-
-/**
- * Loop through every token on uniswap, used for search
- */
-async function getAllTokensOnUniswap() {
-  try {
-    let allFound = false
-    let skipCount = 0
-    let tokens = []
-    while (!allFound) {
-      let result = await client.query({
-        query: ALL_TOKENS,
-        variables: {
-          skip: skipCount
-        },
-        fetchPolicy: 'cache-first'
-      })
-      tokens = tokens.concat(result?.data?.tokens)
-      if (result?.data?.tokens?.length < TOKENS_TO_FETCH || tokens.length > TOKENS_TO_FETCH) {
-        allFound = true
-      }
-      skipCount = skipCount += TOKENS_TO_FETCH
-    }
-    return tokens
-  } catch (e) {
-    console.log(e)
-  }
-}
-
-/**
- * Hook that fetches overview data, plus all tokens and pairs for search
- */
-export function useGlobalData() {
-  const [state, { update, updateAllPairsInUniswap, updateAllTokensInUniswap }] = useGlobalDataContext()
-  const [ethPrice, oldEthPrice] = useEthPrice()
-
-  const data = state?.globalData
-
-  useEffect(() => {
-    async function fetchData() {
-      let globalData = await getGlobalData(ethPrice, oldEthPrice)
-      globalData && update(globalData)
-
-      let allPairs = await getAllPairsOnUniswap()
-      updateAllPairsInUniswap(allPairs)
-
-      let allTokens = await getAllTokensOnUniswap()
-      updateAllTokensInUniswap(allTokens)
-    }
-    if (!data && ethPrice && oldEthPrice) {
-      fetchData()
-    }
-  }, [ethPrice, oldEthPrice, update, data, updateAllPairsInUniswap, updateAllTokensInUniswap])
-
-  return data || {}
-}
-
-export function useGlobalChartData() {
-  const [state, { updateChart }] = useGlobalDataContext()
-  const [oldestDateFetch, setOldestDateFetched] = useState()
-  const [activeWindow] = useTimeframe()
-
-  const chartDataDaily = state?.chartData?.daily
-  const chartDataWeekly = state?.chartData?.weekly
-
-  /**
-   * Keep track of oldest date fetched. Used to
-   * limit data fetched until its actually needed.
-   * (dont fetch year long stuff unless year option selected)
-   */
-  useEffect(() => {
-    // based on window, get starttime
-    let startTime = getTimeframe(activeWindow)
-
-    if ((activeWindow && startTime < oldestDateFetch) || !oldestDateFetch) {
-      setOldestDateFetched(startTime)
-    }
-  }, [activeWindow, oldestDateFetch])
-
-  /**
-   * Fetch data if none fetched or older data is needed
-   */
-  useEffect(() => {
-    let stale = false
-    let intervalRef
-
-    async function fetchData() {
-      // historical stuff for chart
-
-      if (!stale) {
-        // let [newChartData, newWeeklyData] = await getChartData(oldestDateFetch)
-        // updateChart(newChartData, newWeeklyData)
-      }
-    }
-    if (oldestDateFetch && !(chartDataDaily && chartDataWeekly)) {
-      fetchData()
-    }
-
-    intervalRef = setInterval(() => {
-      fetchData()
-    }, 15000)
-
-    return () => {
-      stale = true
-      clearInterval(intervalRef)
-    }
-  }, [chartDataDaily, chartDataWeekly, oldestDateFetch, updateChart])
-
-  return [chartDataDaily, chartDataWeekly]
-}
-
-export function useGlobalTransactions() {
-  const [state, { updateTransactions }] = useGlobalDataContext()
-  const transactions = state?.transactions
-  useEffect(() => {
-    async function fetchData() {
-      if (!transactions) {
-        let txns = await getGlobalTransactions()
-        updateTransactions(txns)
-      }
-    }
-    fetchData()
-  }, [updateTransactions, transactions])
-  return transactions
-}
-
-export function useEthPrice() {
-  const [state, { updateEthPrice }] = useGlobalDataContext()
-  const ethPrice = state?.[ETH_PRICE_KEY]
-  const ethPriceOld = state?.['oneDayPrice']
-  useEffect(() => {
-    async function checkForEthPrice() {
-      if (!ethPrice) {
-        let [newPrice, oneDayPrice, priceChange] = await getEthPrice()
-        updateEthPrice(newPrice, oneDayPrice, priceChange)
-      }
-    }
-    checkForEthPrice()
-  }, [ethPrice, updateEthPrice])
-
-  return [ethPrice, ethPriceOld]
-}
-
-export function useAllPairsInUniswap() {
-  const [state] = useGlobalDataContext()
-  let allPairs = state?.allPairs
-
-  return allPairs || []
-}
-
-export function useAllTokensInUniswap() {
-  const [state] = useGlobalDataContext()
-  let allTokens = state?.allTokens
-
-  return allTokens || []
-}
-
-/**
- * Get the top liquidity positions based on USD size
- * @TODO Not a perfect lookup needs improvement
- */
-export function useTopLps() {
-  const [state, { updateTopLps }] = useGlobalDataContext()
-  let topLps = state?.topLps
-
-  const allPairs = useAllPairData()
-
-  useEffect(() => {
-    async function fetchData() {
-      // get top 20 by reserves
-      let topPairs = Object.keys(allPairs)
-        ?.sort((a, b) => parseFloat(allPairs[a].reserveUSD > allPairs[b].reserveUSD ? -1 : 1))
-        ?.slice(0, 99)
-        .map(pair => pair)
-
-      let topLpLists = await Promise.all(
-        topPairs.map(async pair => {
-          // for each one, fetch top LPs
-          try {
-            const { data: results } = await client.query({
-              query: TOP_LPS_PER_PAIRS,
-              variables: {
-                pair: pair.toString()
-              },
-              fetchPolicy: 'cache-first'
-            })
-            if (results) {
-              return results.liquidityPositions
-            }
-          } catch (e) {}
-        })
-      )
-
-      // get the top lps from the results formatted
-      const topLps = []
-      topLpLists
-        .filter(i => !!i) // check for ones not fetched correctly
-        .map(list => {
-          return list.map(entry => {
-            const pairData = allPairs[entry.pair.id]
-            return topLps.push({
-              user: entry.user,
-              pairName: pairData.token0.symbol + '-' + pairData.token1.symbol,
-              pairAddress: entry.pair.id,
-              token0: pairData.token0.id,
-              token1: pairData.token1.id,
-              usd:
-                (parseFloat(entry.liquidityTokenBalance) / parseFloat(pairData.totalSupply)) *
-                parseFloat(pairData.reserveUSD)
-            })
-          })
-        })
-
-      const sorted = topLps.sort((a, b) => (a.usd > b.usd ? -1 : 1))
-      const shorter = sorted.splice(0, 100)
-      updateTopLps(shorter)
-    }
-
-    if (!topLps && allPairs && Object.keys(allPairs).length > 0) {
-      fetchData()
-    }
-  })
-
-  return topLps
-}
