+++ src/contexts/LocalStorage.js
@@ -5,14 +5,10 @@ const UNISWAP = 'UNISWAP'
 const VERSION = 'VERSION'
 const CURRENT_VERSION = 0
 const LAST_SAVED = 'LAST_SAVED'
-const DISMISSED_PATHS = 'DISMISSED_PATHS'
-const SAVED_ACCOUNTS = 'SAVED_ACCOUNTS'
-const SAVED_TOKENS = 'SAVED_TOKENS'
-const SAVED_PAIRS = 'SAVED_PAIRS'
 
 const DARK_MODE = 'DARK_MODE'
 
-const UPDATABLE_KEYS = [DARK_MODE, DISMISSED_PATHS, SAVED_ACCOUNTS, SAVED_PAIRS, SAVED_TOKENS]
+const UPDATABLE_KEYS = [DARK_MODE]
 
 const UPDATE_KEY = 'UPDATE_KEY'
 
@@ -44,11 +40,7 @@ function reducer(state, { type, payload }) {
 function init() {
   const defaultLocalStorage = {
     [VERSION]: CURRENT_VERSION,
-    [DARK_MODE]: true,
-    [DISMISSED_PATHS]: {},
-    [SAVED_ACCOUNTS]: [],
-    [SAVED_TOKENS]: {},
-    [SAVED_PAIRS]: {}
+    [DARK_MODE]: true
   }
 
   try {
@@ -98,84 +90,4 @@ export function useDarkModeManager() {
     [updateKey, isDarkMode]
   )
   return [isDarkMode, toggleDarkMode]
-}
-
-export function usePathDismissed(path) {
-  const [state, { updateKey }] = useLocalStorageContext()
-  const pathDismissed = state?.[DISMISSED_PATHS]?.[path]
-  function dismiss() {
-    let newPaths = state?.[DISMISSED_PATHS]
-    newPaths[path] = true
-    updateKey(DISMISSED_PATHS, newPaths)
-  }
-
-  return [pathDismissed, dismiss]
-}
-
-export function useSavedAccounts() {
-  const [state, { updateKey }] = useLocalStorageContext()
-  const savedAccounts = state?.[SAVED_ACCOUNTS]
-
-  function addAccount(account) {
-    let newAccounts = state?.[SAVED_ACCOUNTS]
-    newAccounts.push(account)
-    updateKey(SAVED_ACCOUNTS, newAccounts)
-  }
-
-  function removeAccount(account) {
-    let newAccounts = state?.[SAVED_ACCOUNTS]
-    let index = newAccounts.indexOf(account)
-    if (index > -1) {
-      newAccounts.splice(index, 1)
-    }
-    updateKey(SAVED_ACCOUNTS, newAccounts)
-  }
-
-  return [savedAccounts, addAccount, removeAccount]
-}
-
-export function useSavedPairs() {
-  const [state, { updateKey }] = useLocalStorageContext()
-  const savedPairs = state?.[SAVED_PAIRS]
-
-  function addPair(address, token0Address, token1Address, token0Symbol, token1Symbol) {
-    let newList = state?.[SAVED_PAIRS]
-    newList[address] = {
-      address,
-      token0Address,
-      token1Address,
-      token0Symbol,
-      token1Symbol
-    }
-    updateKey(SAVED_PAIRS, newList)
-  }
-
-  function removePair(address) {
-    let newList = state?.[SAVED_PAIRS]
-    newList[address] = null
-    updateKey(SAVED_PAIRS, newList)
-  }
-
-  return [savedPairs, addPair, removePair]
-}
-
-export function useSavedTokens() {
-  const [state, { updateKey }] = useLocalStorageContext()
-  const savedTokens = state?.[SAVED_TOKENS]
-
-  function addToken(address, symbol) {
-    let newList = state?.[SAVED_TOKENS]
-    newList[address] = {
-      symbol
-    }
-    updateKey(SAVED_TOKENS, newList)
-  }
-
-  function removeToken(address) {
-    let newList = state?.[SAVED_TOKENS]
-    newList[address] = null
-    updateKey(SAVED_TOKENS, newList)
-  }
-
-  return [savedTokens, addToken, removeToken]
 }
