+++ src/pages/AccountLookup.js
@@ -7,11 +7,9 @@ import Panel from '../components/Panel'
 import LPList from '../components/LPList'
 import styled from 'styled-components'
 import AccountSearch from '../components/AccountSearch'
-import { useTopLps } from '../contexts/GlobalData'
 import LocalLoader from '../components/LocalLoader'
 import { RowBetween } from '../components/Row'
 import { useMedia } from 'react-use'
-import Search from '../components/Search'
 
 const AccountWrapper = styled.div`
   @media screen and (max-width: 600px) {
@@ -25,7 +23,7 @@ function AccountLookup() {
     window.scrollTo(0, 0)
   }, [])
 
-  const topLps = useTopLps()
+  const topLps = []
 
   const below600 = useMedia('(max-width: 600px)')
 
@@ -34,7 +32,6 @@ function AccountLookup() {
       <FullWrapper>
         <RowBetween>
           <TYPE.largeHeader>Wallet analytics</TYPE.largeHeader>
-          {!below600 && <Search small={true} />}
         </RowBetween>
         <AccountWrapper>
           <AccountSearch />
