+++ src/utils/index.js
@@ -3,8 +3,6 @@ import { BigNumber } from 'bignumber.js'
 import dayjs from 'dayjs'
 import { ethers } from 'ethers'
 import utc from 'dayjs/plugin/utc'
-import { client, blockClient } from '../apollo/client'
-import { GET_BLOCK, GET_BLOCKS, SHARE_VALUE } from '../apollo/queries'
 import { Text } from 'rebass'
 import _Decimal from 'decimal.js-light'
 import toFormat from 'toformat'
@@ -56,34 +54,6 @@ export function getTimeframe(timeWindow) {
   return utcStartTime
 }
 
-export function getPoolLink(token0Address, token1Address = null, remove = false) {
-  if (!token1Address) {
-    return (
-      `https://uniswap.exchange/` +
-      (remove ? `remove` : `add`) +
-      `/${token0Address === '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2' ? 'ETH' : token0Address}/${'ETH'}`
-    )
-  } else {
-    return (
-      `https://uniswap.exchange/` +
-      (remove ? `remove` : `add`) +
-      `/${token0Address === '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2' ? 'ETH' : token0Address}/${
-        token1Address === '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2' ? 'ETH' : token1Address
-      }`
-    )
-  }
-}
-
-export function getSwapLink(token0Address, token1Address = null) {
-  if (!token1Address) {
-    return `https://uniswap.exchange/swap?inputCurrency=${token0Address}`
-  } else {
-    return `https://uniswap.exchange/swap?inputCurrency=${
-      token0Address === '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2' ? 'ETH' : token0Address
-    }&outputCurrency=${token1Address === '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2' ? 'ETH' : token1Address}`
-  }
-}
-
 export function localNumber(val) {
   return Numeral(val).format('0,0')
 }
@@ -101,178 +71,6 @@ export const toWeeklyDate = date => {
   var wkStart = new Date(new Date(date).setDate(date.getDate() - lessDays))
   var wkEnd = new Date(new Date(wkStart).setDate(wkStart.getDate() + 6))
   return dayjs.utc(wkStart).format('MMM DD') + ' - ' + dayjs.utc(wkEnd).format('MMM DD')
-}
-
-export function getTimestampsForChanges() {
-  const utcCurrentTime = dayjs()
-  const t1 = utcCurrentTime
-    .subtract(1, 'day')
-    .startOf('minute')
-    .unix()
-  const t2 = utcCurrentTime
-    .subtract(2, 'day')
-    .startOf('minute')
-    .unix()
-  const tWeek = utcCurrentTime
-    .subtract(1, 'week')
-    .startOf('minute')
-    .unix()
-  return [t1, t2, tWeek]
-}
-
-export async function splitQuery(query, localClient, vars, list, skipCount = 100) {
-  let fetchedData = {}
-  let allFound = false
-  let skip = 0
-
-  while (!allFound) {
-    let end = list.length
-    if (skip + skipCount < list.length) {
-      end = skip + skipCount
-    }
-    let sliced = list.slice(skip, end)
-    let result = await localClient.query({
-      query: query(...vars, sliced),
-      fetchPolicy: 'cache-first'
-    })
-    fetchedData = {
-      ...fetchedData,
-      ...result.data
-    }
-    if (Object.keys(result.data).length < skipCount || skip + skipCount > list.length) {
-      allFound = true
-    } else {
-      skip += skipCount
-    }
-  }
-
-  return fetchedData
-}
-
-/**
- * @notice Fetches first block after a given timestamp
- * @dev Query speed is optimized by limiting to a 600-second period
- * @param {Int} timestamp in seconds
- */
-export async function getBlockFromTimestamp(timestamp) {
-  let result = await blockClient.query({
-    query: GET_BLOCK,
-    variables: {
-      timestampFrom: timestamp,
-      timestampTo: timestamp + 600
-    },
-    fetchPolicy: 'cache-first'
-  })
-  return result?.data?.blocks?.[0]?.number
-}
-
-/**
- * @notice Fetches block objects for an array of timestamps.
- * @dev blocks are returned in chronological order (ASC) regardless of input.
- * @dev blocks are returned at string representations of Int
- * @dev timestamps are returns as they were provided; not the block time.
- * @param {Array} timestamps
- */
-export async function getBlocksFromTimestamps(timestamps, skipCount = 500) {
-  if (timestamps?.length === 0) {
-    return []
-  }
-
-  let fetchedData = await splitQuery(GET_BLOCKS, blockClient, [], timestamps, skipCount)
-
-  let blocks = []
-  if (fetchedData) {
-    for (var t in fetchedData) {
-      if (fetchedData[t].length > 0) {
-        blocks.push({
-          timestamp: t.split('t')[1],
-          number: fetchedData[t][0]['number']
-        })
-      }
-    }
-  }
-  return blocks
-}
-
-export async function getLiquidityTokenBalanceOvertime(account, timestamps) {
-  // get blocks based on timestamps
-  const blocks = await getBlocksFromTimestamps(timestamps)
-
-  // get historical share values with time travel queries
-  let result = await client.query({
-    query: SHARE_VALUE(account, blocks),
-    fetchPolicy: 'cache-first'
-  })
-
-  let values = []
-  for (var row in result?.data) {
-    let timestamp = row.split('t')[1]
-    if (timestamp) {
-      values.push({
-        timestamp,
-        balance: 0
-      })
-    }
-  }
-}
-
-/**
- * @notice Example query using time travel queries
- * @dev TODO - handle scenario where blocks are not available for a timestamps (e.g. current time)
- * @param {String} pairAddress
- * @param {Array} timestamps
- */
-export async function getShareValueOverTime(pairAddress, timestamps) {
-  if (!timestamps) {
-    const utcCurrentTime = dayjs()
-    const utcSevenDaysBack = utcCurrentTime.subtract(8, 'day').unix()
-    timestamps = getTimestampRange(utcSevenDaysBack, 86400, 7)
-  }
-
-  // get blocks based on timestamps
-  const blocks = await getBlocksFromTimestamps(timestamps)
-
-  // get historical share values with time travel queries
-  let result = await client.query({
-    query: SHARE_VALUE(pairAddress, blocks),
-    fetchPolicy: 'cache-first'
-  })
-
-  let values = []
-  for (var row in result?.data) {
-    let timestamp = row.split('t')[1]
-    let sharePriceUsd = parseFloat(result.data[row]?.reserveUSD) / parseFloat(result.data[row]?.totalSupply)
-    if (timestamp) {
-      values.push({
-        timestamp,
-        sharePriceUsd,
-        totalSupply: result.data[row].totalSupply,
-        reserve0: result.data[row].reserve0,
-        reserve1: result.data[row].reserve1,
-        reserveUSD: result.data[row].reserveUSD,
-        token0DerivedETH: result.data[row].token0.derivedETH,
-        token1DerivedETH: result.data[row].token1.derivedETH,
-        roiUsd: values && values[0] ? sharePriceUsd / values[0]['sharePriceUsd'] : 1,
-        ethPrice: 0,
-        token0PriceUSD: 0,
-        token1PriceUSD: 0
-      })
-    }
-  }
-
-  // add eth prices
-  let index = 0
-  for (var brow in result?.data) {
-    let timestamp = brow.split('b')[1]
-    if (timestamp) {
-      values[index].ethPrice = result.data[brow].ethPrice
-      values[index].token0PriceUSD = result.data[brow].ethPrice * values[index].token0DerivedETH
-      values[index].token1PriceUSD = result.data[brow].ethPrice * values[index].token1DerivedETH
-      index += 1
-    }
-  }
-
-  return values
 }
 
 /**
@@ -307,13 +105,6 @@ export const toK = num => {
 export const setThemeColor = theme => document.documentElement.style.setProperty('--c-token', theme || '#333333')
 
 export const Big = number => new BigNumber(number)
-
-export const urls = {
-  showTransaction: tx => `https://etherscan.io/tx/${tx}/`,
-  showAddress: address => `https://www.etherscan.io/address/${address}/`,
-  showToken: address => `https://www.etherscan.io/token/${address}/`,
-  showBlock: block => `https://etherscan.io/block/${block}/`
-}
 
 export const formatDate = unix => {
   const timestamp = dayjs.unix(unix)
