--- src/utils/tokenLists.ts
@@ -1,64 +1,0 @@
-import { TokenList } from '@uniswap/token-lists'
-import schema from '@uniswap/token-lists/src/tokenlist.schema.json'
-import Ajv from 'ajv'
-
-/**
- * Given a URI that may be ipfs, ipns, http, or https protocol, return the fetch-able http(s) URLs for the same content
- * @param uri to convert to fetch-able http url
- */
-function uriToHttp(uri: string): string[] {
-  const protocol = uri.split(':')[0].toLowerCase()
-  switch (protocol) {
-    case 'https':
-      return [uri]
-    case 'http':
-      return ['https' + uri.substr(4), uri]
-    case 'ipfs':
-      const hash = uri.match(/^ipfs:(\/\/)?(.*)$/i)?.[2]
-      return [`https://cloudflare-ipfs.com/ipfs/${hash}/`, `https://ipfs.io/ipfs/${hash}/`]
-    case 'ipns':
-      const name = uri.match(/^ipns:(\/\/)?(.*)$/i)?.[2]
-      return [`https://cloudflare-ipfs.com/ipns/${name}/`, `https://ipfs.io/ipns/${name}/`]
-    default:
-      return []
-  }
-}
-
-const tokenListValidator = new Ajv({ allErrors: true }).compile(schema)
-
-/**
- * Contains the logic for resolving a list URL to a validated token list
- * @param listUrl list url
- */
-export default async function getTokenList(listUrl: string): Promise<TokenList> {
-  const urls = uriToHttp(listUrl)
-  for (let i = 0; i < urls.length; i++) {
-    const url = urls[i]
-    const isLast = i === urls.length - 1
-    let response
-    try {
-      response = await fetch(url)
-    } catch (error) {
-      console.debug('Failed to fetch list', listUrl, error)
-      if (isLast) throw new Error(`Failed to download list ${listUrl}`)
-      continue
-    }
-
-    if (!response.ok) {
-      if (isLast) throw new Error(`Failed to download list ${listUrl}`)
-      continue
-    }
-
-    const json = await response.json()
-    if (!tokenListValidator(json)) {
-      const validationErrors: string =
-        tokenListValidator.errors?.reduce<string>((memo, error) => {
-          const add = `${error.dataPath} ${error.message ?? ''}`
-          return memo.length > 0 ? `${memo}; ${add}` : `${add}`
-        }, '') ?? 'unknown error'
-      throw new Error(`Token list failed validation: ${validationErrors}`)
-    }
-    return json
-  }
-  throw new Error('Unrecognized list URL protocol.')
-}
